--- src/components/advent-of-code/index.tsx
@@ -1,1 +1,0 @@
-export { AdventOfCode } from './components/advent-of-code';
